Revert "SDA-4234: Support Tel Protocol on loading _blank (#1922)"
This reverts commit c56534f3ed12131f1d9f32eaa03ea40cca1c6f3d.
