+++ src/app/window-utils.ts
@@ -54,7 +54,6 @@ import { autoLaunchInstance } from './auto-launch-controller';
 import { autoUpdate, AutoUpdateTrigger } from './auto-update-handler';
 import { mainEvents } from './main-event-handler';
 import { presenceStatus } from './presence-status-handler';
-import { verifyProtocolForNewUrl } from './protocol-handler';
 import { presenceStatusStore } from './stores';
 
 interface IStyles {
@@ -131,7 +130,7 @@ export const preventWindowNavigation = (
   );
 
   const listener = async (e: Electron.Event, winUrl: string) => {
-    if (!verifyProtocolForNewUrl(winUrl)) {
+    if (!winUrl.startsWith('https')) {
       logger.error(
         `window-utils: ${winUrl} doesn't start with https, so, not navigating!`,
       );
